requirements/dev.txt
environs==9.3.2
kafka-python==2.0.2
marshmallow==3.12.2
pymongo==3.12.0
python-dotenv==0.18.0
